add: GreetinsExtension file and remove extension from Greetings class
